+++ Tests/Runtime/FetchConfigsIncomplete_MonobehaviorTest.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using Newtonsoft.Json.Linq;
 using NUnit.Framework;
@@ -7,7 +7,7 @@ using UnityEngine.TestTools;
 
 namespace Unity.RemoteConfig.Tests
 {
-    internal class FetchConfigsComplete_MonobehaviorTest : MonoBehaviour, IMonoBehaviourTest, IRCTest
+    internal class FetchConfigsIncomplete_MonobehaviorTest : MonoBehaviour, IMonoBehaviourTest, IRCTest
     {
         private bool testFinished = false;
         public bool IsTestFinished
@@ -27,7 +27,7 @@ namespace Unity.RemoteConfig.Tests
             responseHeaders.Add("Server", "Jetty(9.4.z-SNAPSHOT)");
 
             var downloadHandlerText =
-                @"{""configs"": {""settings"":{""testInt"":232, ""bool"":true, ""madBro"": ""madAF"", ""someInt"":12, ""heloe"":0.12999999523162842, ""longSomething"": 9223372036854775806, ""stringFormattedAsDate"": ""2020-04-03T10:01:00Z"", ""stringFormattedAsJson"": ""{\""a\"":2.0,\""b\"":4,\""c\"":\""someString\""}"", ""jsonSetting"": ""{\""a\"":1.0,\""b\"":2,\""c\"":\""someString\""}"" }},""metadata"":{""assignmentId"":""3049bfea-05fa-4ddf-acc6-ce43c888fe92"",""environmentId"":""83fff3e2-a945-4601-9ccc-5e9d16d12ea8""}}";
+                @"{""configs"": {""someOtherConfigType"":{""someTestInt"":555}},""metadata"":{""assignmentId"":""5555555-05fa-4ddf-acc6-ce43c888fe92"",""environmentId"":""5555555-a945-4601-9ccc-5e9d16d12ea8""}}";
             var managerImpl = ConfigManager.ConfigManagerImpl;
             var configResponse = managerImpl.ParseResponse(ConfigOrigin.Remote, responseHeaders, downloadHandlerText);
             managerImpl.HandleConfigResponse(ConfigManagerImpl.DefaultConfigKey, configResponse);